status reportes cobranza

+++ routes/web.php
@@ -65,10 +65,10 @@ Route::get('/exportar-hojas-consumos',[DetalleCController::class,'exportPDFGral'
 Route::post('/store-usuarios',[UserController::class,'store'])->name('usuarios.store');
 Route::post('/login',[UserController::class,'create'])->name('usuarios.login');
 Route::post('/logout',[UserController::class,'show'])->name('usuarios.logout');
-Route::get('/administrar-usuarios',[UserController::class,'adminUser'])->name('usuarios.administrar');
-Route::get('/administrar/{id}',[UserController::class,'edit'])->name('usuarios.edit');
-Route::patch('/administrar-modificar/{id}',[UserController::class,'update'])->name('usuarios.update');
-Route::delete('/delete-usuarios/{id}',[UserController::class,'destroy'])->name('usuarios.destroy');
+Route::get('/administrar-usuarios',[UserController::class,'adminUser'])->name('usuarios.administrar')->middleware('auth');
+Route::get('/administrar/{id}',[UserController::class,'edit'])->name('usuarios.edit')->middleware('auth');
+Route::patch('/administrar-modificar/{id}',[UserController::class,'update'])->name('usuarios.update')->middleware('auth');
+Route::delete('/delete-usuarios/{id}',[UserController::class,'destroy'])->name('usuarios.destroy')->middleware('auth');
 
 //Calculo de Comisiones
 Route::get('/comisiones',[ComisionesController::class,'showComisiones'])->name('comisiones.index')->middleware('auth');
